+++ reflex/src/main/java/org/cuba/exception/ReflexException.java
@@ -1,24 +1,48 @@
-package org.cuba.reflex;
+package org.cuba.exception;
 
+/**
+ * @author Kirill Bogatikov
+ * @version 1.0
+ * @since 1.0
+ */
 public class ReflexException extends RuntimeException {
-    private static final long serialVersionUID = -9121480959935548370L;
+    private static final long serialVersionUID = -794179139780556345L;
 
+    /**
+     * 
+     */
     public ReflexException() {
         super();
     }
 
+    /**
+     * @param message
+     */
     public ReflexException(String message) {
         super(message);
     }
 
+    /**
+     * @param cause
+     */
     public ReflexException(Throwable cause) {
         super(cause);
     }
 
+    /**
+     * @param message
+     * @param cause
+     */
     public ReflexException(String message, Throwable cause) {
         super(message, cause);
     }
 
+    /**
+     * @param message
+     * @param cause
+     * @param enableSuppression
+     * @param writableStackTrace
+     */
     public ReflexException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
         super(message, cause, enableSuppression, writableStackTrace);
     }
