+++ reflex/src/main/java/org/cuba/exception/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * @author Kirill Bogatikov
+ * @version 1.0
+ * @since 1.0 *
+ */
+package org.cuba.exception;
